v2/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "asset_monitoring_system"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "pydantic",
    "python-dotenv",
    "httpx",
    "fastapi",
    "uvicorn",
    "sqlmodel",
    "alembic",
    "psycopg2-binary",
    "prefect",
    "pytest-mock",
    "pytest",
    "yfinance",
    "dash",
    "dash-ag-grid",
    "plotly",
    "pandas",
    "dash-bootstrap-components",
    "nbformat>=4.2.0",
    "confluent-kafka",
]

[tool.setuptools.packages.find]
where = ["."]
exclude = ["frontend*", "scripts*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
